refactor: make Countdown a separate component for ConnectionInformation
