+++ src/homepageExperience/components/DataListening.tsx
@@ -10,7 +10,6 @@ import {Button} from '@influxdata/clockface'
 import {
   continuouslyCheckForData,
   TIMEOUT_MILLISECONDS,
-  TIMER_WAIT,
 } from 'src/shared/utils/dataListening'
 
 interface OwnProps {
@@ -80,7 +79,8 @@ class DataListening extends PureComponent<Props, State> {
     return (
       <div className="wizard-step--body-streaming" data-testid="streaming">
         {this.connectionInfo}
-        {this.state.loading === LoadingState.NotFound && (
+        {(this.state.loading === LoadingState.NotFound ||
+          this.state.loading === LoadingState.Error) && (
           <Button onClick={this.handleRetry} text="Retry" />
         )}
       </div>
@@ -116,27 +116,7 @@ class DataListening extends PureComponent<Props, State> {
     } = this.props
 
     this.setState({loading: LoadingState.Loading})
-    this.startTimer()
     continuouslyCheckForData(orgID, bucket, this.updateResponse)
-  }
-
-  private startTimer() {
-    this.setState({timePassedInSeconds: 0, secondsLeft: this.TIMEOUT_SECONDS})
-
-    this.timer = setInterval(this.countDown, TIMER_WAIT)
-  }
-
-  private countDown = () => {
-    const {secondsLeft} = this.state
-    const secs = secondsLeft - 1
-    this.setState({
-      timePassedInSeconds: this.TIMEOUT_SECONDS - secs,
-      secondsLeft: secs,
-    })
-
-    if (secs === 0) {
-      clearInterval(this.timer)
-    }
   }
 }
